Move vim mode line to line 2
Emacs does not allow you to configure the number of lines to read for
the modeline, so Vim's had to be moved instead.
To make the Emacs modeline work, I had to add this in my emacs
configuration:
```elisp
(defalias 'TOML-mode 'conf-toml-mode)
```

To determine the issue with GitHub's syntax highlighting, I looked at
the pull request[^1] that added the emacs/vim modeline detection to
GitHub's syntax highlighter.

[1]: https://github.com/github/linguist/pull/2041