Fix bug where layer references fail inner lookups
The issue was that when using the NodeIndex return type for
lookup_button, buttons pointed to by references were not getting
compiled with the context set to the reference layer, and as a result
internal lookups were failing. To be honest I don't know why I made any
of these decisions when I was writing the code initially, and it took me
a whole hour to figure out enough to fix this bug. This bug was showing
non-deterministic results because in the case where the button itself
was compiled before the reference, there would be no issue as
lookup_button resolution would be done with the correct context (the
layer containing the button itself), but when it was being compiled
/from/ the reference things would break. What makes this
non-deterministic is that the order of compilation depends on
*HashMap iteration order*, which changes run-by-run.