Change specified example filetypes from TOML to INI
ini renders a lot better on GitHub for our language; mainly, it doesn't
have all the red highlights. One unfortunate consequence is that because
Emacs doesn't have an `ini-mode` (instead calling it `conf-mode`), I had
to delete the Emacs mode line altogether, as otherwise it would cause
GitHub to render everything as a `conf` file, which does not have great
highlighting.