fix: override keys from initial inherited layers
When processing a new layer, its first parent layer was being treated
specially, in that the new layer was created as a verbatim copy of the
parent layer; this meant that when setting the first inherited layer to
a non-builtin layers, any of its explicitly defined keys would /not/
have `is_inherited` set to true. This meant that for the rest of the
inheritance process, those keys would appear to be defined inline in the
new layer, causing them to take precedence over any future parent layer
keys.